+++ tests/benchdnn/common.cpp
@@ -387,9 +387,10 @@ void array_set(char *arr, size_t size) {
 }
 
 void gemm(const char *layout, const char *transa, const char *transb,
-        int m, int n, int k, const float alpha, const float *a, const int lda,
-        const float *b, const int ldb, const float beta, float *c,
-        const int ldc ) {
+        int64_t m, int64_t n, int64_t k,
+        const float alpha, const float *a, const int64_t lda,
+        const float *b, const int64_t ldb,
+        const float beta, float *c, const int64_t ldc) {
     if (*layout == 'F') {
         mkldnn_sgemm(transa, transb, &m, &n, &k, &alpha, a, &lda, b, &ldb,
                 &beta, c, &ldc);
